WIP: asrock x570d4u
This commit adds meta layer support for asrock rack X570D4U motherboard.
This board contains an ast2500 BMC, an AMD am4 (ryzen) CPU.  Both SPI
flashes are socketed from the factory, giving easy access to flash
OpenBMC on it with a flash programmer.

Specifically, this image uses the UBI filesystem, which should give some
advantages in terms of utilization of flash space on the BMC.

This patchset is WIP, as the devicetree needs to be sent to LKML.

Change-Id: Idbb5416bfe96a17386b962b0a358c34490254a1c