WIP:Separate UBI support and dual flash support
Ideally UBI support in an image should be separable from whether or not
a platform has dual flash chips onboard.  This patchset attempts to
split apart the feature along those lines, and allow UBI images to be
selected separately from dual flash images via which distro feature gets
included.

phosphor-ubi-df.inc is changed to be the image with dual flash
phosphor-ubi.inc generates an ubi image for a single flash.

Mechanically, this involved splitting
config-ast-common-Add-bootopts-to-support-ubi-and-mt into two patches.

witherspoon appears to be the only user of phosphor-ubi, and seems to
contain dual flash, so it mechanically moves witherspoon to the new
distro feature.

This patchset is a WIP.  It needs more testing.
Tested-by: Adriana Kobylak <anoo@us.ibm.com>
Tested: Flashed the generated ubi.mtd file to a Witherspoon and verified
        it booted and that the ubi volumes and u-boot environment looked
        as expected.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I295028bef505d015f17d725bd1b3985148e20fb0